# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Exoplanet Hunter API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PredictInput {
  /** Flux values from the light curve (3197 values expected) */
  flux: number[];
  /**
     * Optional label for the sample (for display purposes)
     * @nullable
     */
  label?: string | null;
}

/**
 * Probabilities for each class
 */
export type PredictResultProbabilities = {[key: string]: number};

export interface PredictResult {
  /** Predicted label: EXOPLANET or NON-EXOPLANET */
  label: string;
  /** Confidence of the prediction (0-1) */
  confidence: number;
  /** Whether the star is predicted to host an exoplanet */
  is_exoplanet: boolean;
  /** Probabilities for each class */
  probabilities: PredictResultProbabilities;
}

export interface SampleLightCurve {
  id: string;
  /** EXOPLANET or NON-EXOPLANET */
  label: string;
  flux: number[];
  description: string;
  /** @nullable */
  star_id?: string | null;
}

export interface ModelStatus {
  ready: boolean;
  message: string;
}

export interface ApiError {
  error: string;
}