# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Exoplanet Hunter API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /predict:
    post:
      operationId: predict
      tags: [model]
      summary: Predict exoplanet
      description: Classify a light curve as exoplanet or non-exoplanet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PredictInput"
      responses:
        "200":
          description: Prediction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PredictResult"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Model not ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    PredictInput:
      type: object
      required:
        - flux
      properties:
        flux:
          type: array
          items:
            type: number
          description: Flux values from the light curve (3197 values expected)
        label:
          type: ["string", "null"]
          description: Optional label for the sample (for display purposes)
    PredictResult:
      type: object
      required:
        - label
        - confidence
        - is_exoplanet
        - probabilities
      properties:
        label:
          type: string
          description: "Predicted label: EXOPLANET or NON-EXOPLANET"
        confidence:
          type: number
          description: Confidence of the prediction (0-1)
        is_exoplanet:
          type: boolean
          description: Whether the star is predicted to host an exoplanet
        probabilities:
          type: object
          description: Probabilities for each class
          additionalProperties:
            type: number
...
    ApiError:
      type: object
      required:
        - error
      properties:
        error:
          type: string