# lib/api-zod/src/generated/types/predictInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Exoplanet Hunter API
 * OpenAPI spec version: 0.1.0
 */

export interface PredictInput {
  /** Flux values from the light curve (3197 values expected) */
  flux: number[];
  /**
     * Optional label for the sample (for display purposes)
     * @nullable
     */
  label?: string | null;
}