# lib/api-zod/src/generated/types/predictResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Exoplanet Hunter API
 * OpenAPI spec version: 0.1.0
 */
import type { PredictResultProbabilities } from './predictResultProbabilities';

export interface PredictResult {
  /** Predicted label: EXOPLANET or NON-EXOPLANET */
  label: string;
  /** Confidence of the prediction (0-1) */
  confidence: number;
  /** Whether the star is predicted to host an exoplanet */
  is_exoplanet: boolean;
  /** Probabilities for each class */
  probabilities: PredictResultProbabilities;
}